close menu on selection and encapsulate into custom hook
