+++ src/functions/hooks/index.ts
@@ -4,3 +4,4 @@ export { useReceiveDashboard } from "./useReceiveDashboard";
 export { useInitDashboards } from "./useInitDashboards";
 export { useFixLayout } from "./useFixLayout";
 export { useSyncStorage } from "./useSyncStorage";
+export { useDashboardSearch } from "./useDashboardSearch";
