+++ src/functions/hooks/useDashboardSearch.ts
@@ -1,0 +1,67 @@
+import OBR, { type Image } from "@owlbear-rodeo/sdk";
+import { useEffect } from "react";
+
+import { useAppStore } from "../../AppProvider";
+import db from "../../dbInstance";
+
+export const useDashboardSearch = () => {
+  const { selectedDashboard, selectADashboard } = useAppStore();
+
+  useEffect(() => {
+    const setupContextMenu = () => {
+      OBR.contextMenu.create({
+        id: `com.roberttate.dashboard-maker-search-context-menu`,
+        icons: [
+          {
+            icon: "/icon.svg",
+            label: "Dashboard Search",
+            filter: {
+              every: [{ key: "type", value: "IMAGE" }],
+            },
+          },
+        ],
+        onClick: async (context) => {
+          const isImageItemType = context?.items?.[0]?.type === "IMAGE";
+          if (isImageItemType) {
+            const imageItem = context?.items?.[0] as Image;
+            const givenTokenName = imageItem?.text?.plainText || "";
+            const keys = await db.keys();
+            if (givenTokenName === selectedDashboard) {
+              await OBR.player.deselect();
+              await OBR.notification.show(
+                `That dashboard is already open!`,
+                "SUCCESS",
+              );
+            } else if (keys.includes(givenTokenName)) {
+              await OBR.player.deselect();
+              selectADashboard(givenTokenName);
+              await OBR.action.open();
+            } else if (givenTokenName === "") {
+              await OBR.player.deselect();
+              await OBR.notification.show(
+                `No Dashboard found. Give this token a name that matches a dashboard first!`,
+                "ERROR",
+              );
+            } else {
+              await OBR.player.deselect();
+              await OBR.notification.show(
+                `No Dashboard found by that name.`,
+                "ERROR",
+              );
+            }
+          } else {
+            await OBR.player.deselect();
+            await OBR.notification.show(
+              `That token type doesn't work for dashboard searching. Sorry!`,
+              "ERROR",
+            );
+          }
+        },
+      });
+    };
+
+    OBR.onReady(() => {
+      setupContextMenu();
+    });
+  }, [selectedDashboard]);
+};
